Use link for XDP
